+++ include/chipmunk/chipmunk_structs.h
@@ -87,7 +87,7 @@ enum cpArbiterState {
 	// Arbiter is active and its not the first collision.
 	CP_ARBITER_STATE_NORMAL,
 	// Collision has been explicitly ignored.
-	// Either by returning false from a begin collision handler or calling cpArbiterIgnore().
+	// Can be set/unset from begin or preStep callback functions.
 	CP_ARBITER_STATE_IGNORE,
 	// Collison is no longer active. A space will cache an arbiter for up to cpSpace.collisionPersistence more steps.
 	CP_ARBITER_STATE_CACHED,
@@ -143,6 +143,7 @@ struct cpArbiter {
 	
 	cpTimestamp stamp;
 	enum cpArbiterState state;
+	cpBool processCollision;
 };
 
 struct cpShapeMassInfo {
@@ -431,9 +432,8 @@ struct cpSpace {
 	cpArray *allocatedBuffers;
 	int locked;
 	
-	cpBool usesWildcards;
 	cpHashSet *collisionHandlers;
-	cpCollisionHandler defaultHandler;
+	cpCollisionHandler globalHandler;
 	
 	cpBool skipPostStep;
 	cpArray *postStepCallbacks;