+++ include/chipmunk/cpArbiter.h
@@ -58,9 +58,16 @@ CP_EXPORT cpVect cpArbiterTotalImpulse(const cpArbiter *arb);
 /// This function should only be called from a post-solve, post-step or cpBodyEachArbiter callback.
 CP_EXPORT cpFloat cpArbiterTotalKE(const cpArbiter *arb);
 
-/// Mark a collision pair to be ignored until the two objects separate.
-/// Pre-solve and post-solve callbacks will not be called, but the separate callback will be called.
-CP_EXPORT cpBool cpArbiterIgnore(cpArbiter *arb);
+/// Return if the collision was set to be ignored by cpArbiterSetProcessCollision earlier.
+/// Note that the collision can be ignored for different reasons, even when this is true.
+CP_EXPORT cpBool cpArbiterGetProcessCollision(const cpArbiter *arb);
+/// Set processCollision to false to not process the collision. 
+/// Setting this to false from a begin callback causes the collision to be ignored until
+/// the the separate callback is called when the objects stop colliding. 
+/// Returning false from a pre-step callback causes the collision to be ignored until the next step.
+/// It should not be called from post-solve or separate callbacks.
+CP_EXPORT void cpArbiterSetProcessCollision(cpArbiter *arb, cpBool processCollision);
+
 
 /// Return the colliding shapes involved for this arbiter.
 /// The order of their cpSpace.collision_type values will match
@@ -118,29 +125,5 @@ CP_EXPORT cpVect cpArbiterGetPointA(const cpArbiter *arb, int i);
 CP_EXPORT cpVect cpArbiterGetPointB(const cpArbiter *arb, int i);
 /// Get the depth of the @c ith contact point.
 CP_EXPORT cpFloat cpArbiterGetDepth(const cpArbiter *arb, int i);
-
-/// If you want a custom callback to invoke the wildcard callback for the first collision type, you must call this function explicitly.
-/// You must decide how to handle the wildcard's return value since it may disagree with the other wildcard handler's return value or your own.
-CP_EXPORT cpBool cpArbiterCallWildcardBeginA(cpArbiter *arb, cpSpace *space);
-/// If you want a custom callback to invoke the wildcard callback for the second collision type, you must call this function explicitly.
-/// You must decide how to handle the wildcard's return value since it may disagree with the other wildcard handler's return value or your own.
-CP_EXPORT cpBool cpArbiterCallWildcardBeginB(cpArbiter *arb, cpSpace *space);
-
-/// If you want a custom callback to invoke the wildcard callback for the first collision type, you must call this function explicitly.
-/// You must decide how to handle the wildcard's return value since it may disagree with the other wildcard handler's return value or your own.
-CP_EXPORT cpBool cpArbiterCallWildcardPreSolveA(cpArbiter *arb, cpSpace *space);
-/// If you want a custom callback to invoke the wildcard callback for the second collision type, you must call this function explicitly.
-/// You must decide how to handle the wildcard's return value since it may disagree with the other wildcard handler's return value or your own.
-CP_EXPORT cpBool cpArbiterCallWildcardPreSolveB(cpArbiter *arb, cpSpace *space);
-
-/// If you want a custom callback to invoke the wildcard callback for the first collision type, you must call this function explicitly.
-CP_EXPORT void cpArbiterCallWildcardPostSolveA(cpArbiter *arb, cpSpace *space);
-/// If you want a custom callback to invoke the wildcard callback for the second collision type, you must call this function explicitly.
-CP_EXPORT void cpArbiterCallWildcardPostSolveB(cpArbiter *arb, cpSpace *space);
-
-/// If you want a custom callback to invoke the wildcard callback for the first collision type, you must call this function explicitly.
-CP_EXPORT void cpArbiterCallWildcardSeparateA(cpArbiter *arb, cpSpace *space);
-/// If you want a custom callback to invoke the wildcard callback for the second collision type, you must call this function explicitly.
-CP_EXPORT void cpArbiterCallWildcardSeparateB(cpArbiter *arb, cpSpace *space);
 
 /// @}