+++ src/cpSpaceStep.c
@@ -253,19 +253,30 @@ cpSpaceCollideShapes(cpShape *a, cpShape *b, cpCollisionID id, cpSpace *space)
 	cpArbiterUpdate(arb, &info, space);
 	
 	cpCollisionHandler *handler = arb->handler;
+	cpCollisionHandler *handlerA = arb->handlerA;
+	cpCollisionHandler *handlerB = arb->handlerB;
+	cpCollisionHandler *handlerGlobal = &space->globalHandler;
 	
 	// Call the begin function first if it's the first step
-	if(arb->state == CP_ARBITER_STATE_FIRST_COLLISION && !handler->beginFunc(arb, space, handler->userData)){
-		cpArbiterIgnore(arb); // permanently ignore the collision until separation
+	if(arb->state == CP_ARBITER_STATE_FIRST_COLLISION) {
+		handler->beginFunc(arb, space, handler->userData);
+		handlerA->beginFunc(arb, space, handlerA->userData);
+		arb->swapped = !arb->swapped;
+		handlerB->beginFunc(arb, space, handlerB->userData);
+		arb->swapped = !arb->swapped;
+		handlerGlobal->beginFunc(arb, space, handlerGlobal->userData);
 	}
-	
+
+	// Call the preSolve function
+	handler->preSolveFunc(arb, space, handler->userData);
+	handlerA->preSolveFunc(arb, space, handlerA->userData);
+	arb->swapped = !arb->swapped;
+	handlerB->preSolveFunc(arb, space, handlerB->userData);
+	arb->swapped = !arb->swapped;
+	handlerGlobal->preSolveFunc(arb, space, handlerGlobal->userData);
 	if(
-		// Ignore the arbiter if it has been flagged
-		(arb->state != CP_ARBITER_STATE_IGNORE) && 
-		// Call preSolve
-		handler->preSolveFunc(arb, space, handler->userData) &&
-		// Check (again) in case the pre-solve() callback called cpArbiterIgnored().
-		arb->state != CP_ARBITER_STATE_IGNORE &&
+		// Ignore the arbiter if it has been flagged from the begin or preSolve funcs.
+		(arb->state != CP_ARBITER_STATE_IGNORE) &&
 		// Process, but don't add collisions for sensors.
 		!(a->sensor || b->sensor) &&
 		// Don't process collisions between two infinite mass bodies.
@@ -311,7 +322,15 @@ cpSpaceArbiterSetFilter(cpArbiter *arb, cpSpace *space)
 	if(ticks >= 1 && arb->state != CP_ARBITER_STATE_CACHED){
 		arb->state = CP_ARBITER_STATE_CACHED;
 		cpCollisionHandler *handler = arb->handler;
+		cpCollisionHandler *handlerA = arb->handlerA;
+		cpCollisionHandler *handlerB = arb->handlerB;
+		cpCollisionHandler *handlerGlobal = &space->globalHandler;
 		handler->separateFunc(arb, space, handler->userData);
+		handlerA->separateFunc(arb, space, handlerA->userData);
+		arb->swapped = !arb->swapped;
+		handlerB->separateFunc(arb, space, handlerB->userData);
+		arb->swapped = !arb->swapped;
+		handlerGlobal->separateFunc(arb, space, handlerGlobal->userData);
 	}
 	
 	if(ticks >= space->collisionPersistence){
@@ -440,7 +459,15 @@ cpSpaceStep(cpSpace *space, cpFloat dt)
 			cpArbiter *arb = (cpArbiter *) arbiters->arr[i];
 			
 			cpCollisionHandler *handler = arb->handler;
+			cpCollisionHandler *handlerA = arb->handlerA;
+			cpCollisionHandler *handlerB = arb->handlerB;
+			cpCollisionHandler *handlerGlobal = &space->globalHandler;
 			handler->postSolveFunc(arb, space, handler->userData);
+			handlerA->postSolveFunc(arb, space, handlerA->userData);
+			arb->swapped = !arb->swapped;
+			handlerB->postSolveFunc(arb, space, handlerB->userData);
+			arb->swapped = !arb->swapped;
+			handlerGlobal->postSolveFunc(arb, space, handlerGlobal->userData);
 		}
 	} cpSpaceUnlock(space, cpTrue);
 }